Proof-of-concept CreateTweet Cloudflare bypass
Grab binaries from https://github.com/lexiforest/curl-impersonate and drop
them in the bin folder. This is currently messy and non-portable. Posting
tweets does work, but the frontend will not acknowledge them due to the
response body of requests currently being garbage.

This will all be rewritten in the future.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>